Initial Commit
Setup the testing for the project.

I'm going to try implementing my own event handling API on top of the Gamepad API now, the ones I've been trying to use haven't been meeting my needs just yet. I want to have something that keeps a solid reference to the state, and gets events based on when the data changes. Then it can work nicely with reactive frameworks, UI handling, and state management, like for Flatlands, Minecraft, NBT structures, SolidJS, React, things like that. Not that those are new concerns, but I have a picture of how I want it to work, and it doesn't seem like other projects have implemented it like that yet.