Overloads are King!!! 👑
That was the issue with the declaration of the fields. The built-in DOM interfaces also use this overload-based syntax, so while it's not ideal either way to redeclare things in that way, it seems like that's how the rest of the standard library is doing it, so it isn't something to be worried about in terms of my ability of implementing this. If they change the standard library from implementing it like this as well, then lots of more significant changes will happen before this small project needs to adjust itself as well.