Dot Property Revert
Going to maybe add these later on, I want to focus on implementing things nicely before I add niceties to the API itself. I want this to be very vanilla-feeling to the web platform itself, so I want to add each part only as I know it works well! :D