Polling Events
I want to have a singular 'game loop' that runs for checking any instances of the state handler, so it should be very performant, as I don't want to have any race conditions, or slowdowns because of running multiple loops in the background. Ideally you'd have only a singular loop for your game and the controller polling, but I think having to tie those two together is also a bit of a headache. I'm not sure that doing that will help prevent any performance concerns either, they might not be there to begin with.