Input Event!!!!!
Wow, can't believe this actually works! It sends an update event for anytime any of the controller inputs change. This makes it a bunch easier to respond to, I'm curious about how this can work with the UI as well.

I didn't know until just this session that the timestamp only changes when one of the inputs changes, that's an exceptionally helpful thing to know now, I was worried I was going to have to diff out the entire tree of input values, thankfully that doesn't seem to be the case.

I still have the issue though where my SteelSeries Nimus doesn't have a proper mapping, and it shows up as two distinct controllers, which is a bit annoying. I wonder if I can write an extension to fix it.

So the cool part about this is that you can have multiple state hooks for a single controller, and each controller get's it's own event stack. So you just pass in which index you want to use, and that's what gets the validation for you! No need to sort through the list on your own. I'm curious if there's a better way I should separate this state out of the class though, maybe with moving the loop outside into it's own class or function, rather than using static methods. But hey, it's type-safe, and it's working somehow! Getting closer with each little step. Slow progress, is still progress.