Gamepad Logging Retest
Coming back to check this project out again

SYL - The New Black

Looking into if this properly works like how I think it does

So the goal is to only send events around while the single loop is running in the background, and only notify the specific sender when one of it's values changes. It should be very performant, because it only has a single loop running, and the events are only dispatched when the controller itself changes.