Updates Display
I should make a proper setup to render this, it helps make it more understandable much better.

The thing here is to show when each of the controllers are updating, and that each handler is specific to that controller index. My Nimbus controller for some reason is represented by two controllers, so that's what I'm trying to debug here. In my games that use the Gamepad API, I'll have to add a controller picker for browsers like Chrome that don't handle the controller completely right. Safari seems to be okay with it though, so it must be the macOS implementation in Chromium that doesn't work completely with MFi devices.