Observer Pattern Exploration
This was leftover code from the session 3 days ago, I don't think it's complete, and I don't remember if it's functional. Coming back to it today though, I realized it makes sense to document it as a concept I was looking into, since I think I will be moving towards this, away from the original events setup.