Next Observer Demo
The rehash concept, I started building it from scratch, this one isn't complete either. I'm looking to that of `IntersectionObserver` and `MutationObserver` and have found that the pattern for those just might be how I want to handle things here. Been looking at a bunch of StackOverflow posts around the two, TypeScript types, and usage patterns.