Multi-Record Preparations
Realized that for the other observer APIs, they batch their callback calls together into a single call, and the results are returned in an array of the current changed entries, rather than calling each update on it's own. Now I just have to update the calling and handling to batch these results together on my end as well. The API is shaped accordingly now, it just doesn't do that kind of thing functionally-speaking yet.

Also, should my record object be a class of some kind? That seems to be what the vanilla APIs do, but I'm not completely sure I see why it has to be a class, rather than just a plain object. Maybe just consistency, and for the possibility of providing methods to call on the record object, be it needing a feature like that?