Observer Start/Stop Handling
So I need to figure out the grouping for records still. Other than that, I tried re-adding the start and stop detection for the internal loop, I don't want to do it with events, I thought I'd try it just to see if it works nicely though. It might not be necessary to turn this internal loop on and off all the time as controllers connect and disconnect, is it? I might worry about that later on, in favor of making other API things work before worrying about that aspect of possible performance worries. This is the same kind of recursive loop you have to do in your own apps right now anyways, so it's not like it's something new that we aren't doing already, it's just abstracted away into this observer pattern.

Wanted to try this because I wanted to make sure the callback scenarios get called in the correct order as to how it should make sense.