Loop De-Dupe Test
Trying to make the loop only run when a controller is connected, and to pause it when there aren't any. I also have to keep in mind that controllers can possibly join at one point or another, and I think it's possibly for a controller to slip through or out, if the callbacks aren't run in the right order and timing.