Solid Demo Setup
Going to use Solid for my demo to help with testing things out faster, in terms of not needing to write out all of the UI logic by hand for the whole thing. I want to be able to easily test whether my code is working, and spend less headspace on getting the UI for that up and running. This is just for the dev side of things, just my own convenience. Could be helpful to dogfood the API in this way too, to see if it works as nicely as intended inside functional frameworks too :D