Poll Event Typings
Made this more similar to regular event callback typings. I was going to look into extending `EventTarget` for these again, and still keep the observer paradigm, but it feels weird to extend that just for that single use-case, as well as to use both concepts in a single class. Maybe it's not that weird? It seems kind of weird. I think it's just because the internal polling loop is something different than the observables themselves. Maybe I can just make the loop module-scoped and global for all observers? That's kind of what I was doing before too? I'll look into that next.