Debounce Pool Demo
Trying to figure out how to group records that are created between each frame, rather than firing calls for each one individually.

I tried to combine a pool setup with that of a debounce/throttle kind of thing, based around `requestAnimationFrame()` too. It doesn't seem to be working completely as expected, as the debounced entries seem to be lost?

https://gist.github.com/rpemberton/b15a206f453ff233830f3183e4ac371a

DTP Ki, and my own songs Safe Place, and Sound Bath.