Poll Animation Timing Adjustment
I think this might help fix the timing around the debug issue. I think I have to adjust the debounce function to line up with the polling loop itself, I think the timings were weird, in that it was clearing out the pool before the frame was able to send the records to the callback.

So in other words, I think I just have to return a Promise from this, which resolves with each frame call. Maybe I should just add a poll callback instead? That might just solve it too, come to think of it.