Poll Grouping/Timing
Trying to recount how the timings of these should work. Right now the events are firing in the wrong order with this setup, not sure why yet. This fixes the grouping though, in that these should all happen in the same frame. I want the events to fire alongside the timer of the loop itself, rather than randomly as the happen. Or *should* they be as fast as possible? Maybe only `input` should fire on the clock? ahah, might be on to something!