Further Grouping Demo (flop.?)
Not quite working as expected, still, again aagh! Maybe I should just to the single callback system, rather than grouping things into an array? I wanted to make it similar to that of the other observer APIs, but maybe it doesn't make complete sense to do that with the results for these. I think it does make a lot of sense for using the observing pattern for each of the controller indexes though, that works great.

(flp.)