Single Record Callback
Since the timings of these don't appear to line up exactly, I'm going to make the callback fire independently for each gamepad update, and based on what the record type is.

Need to test this in other browsers now, and ensure that the timings of everything are correct, and smoothen out any API weirdness and implementation intricacies.