Object Refs Fix; Timestamps!
Turns out (from what I think, need to investigate more) that Chrome's return value for `navigator.getGamepads()` returns new Gamepad instances with each call (this part I knew), but Firefox and Safari actually return the same object reference itself! So storing the gamepads in a record doesn't work to diff out the timestamps, because those values themselves are changed, hence meaning they can never pass as being different, because the one that was stored is the same one that gets returned from the Navigator method, they are 'both changed', when in reality they are the same reference under the hood.

I was bummed that this wasn't working, it's always the darndest of funky things, huh??

So my fix for this is to save just the timestamps on their own, since those are always pass-by-value. I wasn't using any of the other info on those objects either, so it was more bloated than it needed to be.