TSC Configs (Mostly for build)
Now that I've tested this out on multiple browsers, I think I'm gonna get ready to package it up for usage on npm!

Noticed in the previous commit that the precision of `performance.now()` being low in Firefox makes the diff loop much slower, because the content of the Gamepads update, but the timestamp didn't change yet, because it doesn't show the smaller portion of the timestamp.

#1