Disposable for All Classes!
At first I was going to remove the disposable feature since it didn't initially seem necessary to me, but maybe it could be a nice shortcut to get a simple set of information, like if you wanted to wait for a specific gamepad change in a Promise hook or something. It could be nice to just be able to use `using` in the new TypeScript ponyfill, say if it does make your life easier for a random thing. Forgetting to call `disconnect()` or `dispose()` for one or the other of the classes, it wouldn't do much for this, but if you did it a bunch of times over, I think it would be keen to make it easy for users to not need to keep track of their state of it, if you can let the language handle it for you out of the box. Might as well provide the option, rather than thinking it's one way or the other?