Reactive Support, Conceptually..?
Realized that now this should be simple to support in terms of having auto-cleanup for reactive component frameworks.

With this I also discovered that my state handling doesn't account for when you create new observers after your gamepads have already been connected. They should be able to be instantiated whenever, and know how to take the current state of the page and start working with them from there, getting info on the gamepads onwards from whenever you made the watcher.