Queue the Microtasks!
Wow, this is a very awesome tool! I kind of brainstormed out of some part of my brain that maybe this is what I was looking for here. It seems to work just right!

The issue was that the constructor was trying to start the detection code before anyone would have any chance of setting any event listeners on the observer itself, so they wouldn't properly get fired for when the gamepads were 'connected', per se.

https://macarthur.me/posts/navigating-the-event-loop/